# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MealPlan+ API
  description: |
    REST API for MealPlan+ — nutrition tracking, meal planning & fitness.

    ## Authentication
    All endpoints require a Firebase ID token in the `Authorization: Bearer <token>` header.
    The backend resolves the token to a `firebaseUid` and auto-creates the user on first call.

    ## Sync strategy
    Client-server sync uses **last-write-wins on `updatedAt`**.
    - Every syncable resource carries `serverId` (stable UUID, survives local DB wipes) and `updatedAt`.
    - **Pull:** `GET /api/v1/sync/pull?since=<ISO-8601>` — returns all records changed after `since`.
    - **Push:** `POST /api/v1/sync/push` — upserts records by `serverId`; server returns count accepted.
    - Android `SyncWorker` stores `serverTime` from the previous pull response and passes it as `since`.
    - Soft-deleted records are communicated via `tombstones[]` in the pull response.

    ## Macros
    All food macros are stored and returned **per 100 g**. Clients compute per-serving values
    using `gramsPerPiece` / `gramsPerCup` / `gramsPerTbsp` / `gramsPerTsp` as needed.
  version: v1
servers:
  - url: https://mealplan-api-rfo22lhanq-ez.a.run.app
    description: Production (Cloud Run)
  - url: http://localhost:8080
    description: Local dev
security:
  - FirebaseJWT: []
...
paths:
...
  /api/v1/foods:
    get:
      tags: [Foods]
      summary: List foods
      description: Returns all system foods plus foods created by the authenticated user.
      operationId: listFoods
      parameters:
        - name: favorites
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: When true, returns only items where `isFavorite = true`.
      responses:
        '200':
          description: List of foods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FoodDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Foods]
      summary: Create food
      operationId: createFood
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodDto'
      responses:
        '201':
          description: Created food
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/foods/search:
    get:
      tags: [Foods]
      summary: Search foods
      description: Full-text search across system and user foods by name or brand. Returns a paginated result.
      operationId: searchFoods
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            default: ""
          description: Search query (name or brand)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 30
            maximum: 100
      responses:
        '200':
          description: Paginated food results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/foods/{id}:
    get:
      tags: [Foods]
      summary: Get food by ID
      operationId: getFood
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Food
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Foods]
      summary: Update food
      description: Full replace of a user-owned food. System foods cannot be updated.
      operationId: updateFood
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodDto'
      responses:
        '200':
          description: Updated food
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Foods]
      summary: Delete food
      operationId: deleteFood
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/foods/{id}/favorite:
    patch:
      tags: [Foods]
      summary: Toggle favourite
      description: |
        Toggles `isFavorite` for the food. For system foods, the preference is stored
        per-user in `food_user_prefs`; for user-owned foods it updates `foods.is_favorite`.
      operationId: toggleFoodFavorite
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Food with updated `isFavorite`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Meals ───────────────────────────────────────────────────────────────────
...
  /api/v1/exercises:
    get:
      tags: [Exercises]
      summary: List exercises
      description: Returns system exercises plus user-created exercises. Optionally filter by name or tag.
      operationId: listExercises
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive name search (pushed to SQL ILIKE).
        - name: tagId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter to exercises tagged with this tag ID.
      responses:
        '200':
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExerciseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Exercises]
      summary: Create exercise
      operationId: createExercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseDto'
      responses:
        '201':
          description: Created exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/exercises/{id}:
    get:
      tags: [Exercises]
      summary: Get exercise by ID
      operationId: getExercise
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Exercises]
      summary: Update exercise
      operationId: updateExercise
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseDto'
      responses:
        '200':
          description: Updated exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Exercises]
      summary: Delete exercise
      description: Deletes exercise and cascades tag associations. Removing from workout templates is not automatic — clients must handle orphan cleanup.
      operationId: deleteExercise
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Workout Templates ────────────────────────────────────────────────────────
  /api/v1/workout-templates:
    get:
      tags: [Workout Templates]
      summary: List workout templates
      operationId: listWorkoutTemplates
      responses:
        '200':
          description: List of workout templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkoutTemplateDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Workout Templates]
      summary: Create workout template
      operationId: createWorkoutTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutTemplateDto'
      responses:
        '201':
          description: Created workout template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutTemplateDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workout-templates/{id}:
    get:
      tags: [Workout Templates]
      summary: Get workout template by ID
      operationId: getWorkoutTemplate
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Workout template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutTemplateDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Workout Templates]
      summary: Update workout template
      operationId: updateWorkoutTemplate
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutTemplateDto'
      responses:
        '200':
          description: Updated workout template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutTemplateDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Workout Templates]
      summary: Delete workout template
      operationId: deleteWorkoutTemplate
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/workout-templates/{id}/start:
    post:
      tags: [Workout Templates]
      summary: Start a workout session from template
      description: |
        Creates a new `WorkoutSession` pre-populated with sets from the template
        (one `WorkoutSet` per target set, with `reps` and `weightKg` from the template).
        The returned session is in `isCompleted = false` state — call `/finish` when done.
      operationId: startWorkoutFromTemplate
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '201':
          description: New workout session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSessionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Workout Sessions ─────────────────────────────────────────────────────────
  /api/v1/workout-sessions:
    get:
      tags: [Workout Sessions]
      summary: List workout sessions
      operationId: listWorkoutSessions
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start of date range (inclusive)
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End of date range (inclusive)
      responses:
        '200':
          description: List of workout sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkoutSessionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Workout Sessions]
      summary: Create workout session
      operationId: createWorkoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutSessionDto'
      responses:
        '201':
          description: Created session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSessionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workout-sessions/last-for-exercise/{exerciseId}:
    get:
      tags: [Workout Sessions]
      summary: Get last logged sets for an exercise
      description: |
        Returns the sets from the most recent **completed** session that contains
        the given exercise. Used by the Session Runner "Last time" panel.
      operationId: lastSetsForExercise
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Last sets for exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastSetsDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workout-sessions/{id}:
    get:
      tags: [Workout Sessions]
      summary: Get workout session by ID
      operationId: getWorkoutSession
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Workout session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSessionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Workout Sessions]
      summary: Update workout session
      description: Replaces sets. Use to log actual reps/weight during an active session.
      operationId: updateWorkoutSession
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutSessionDto'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSessionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Workout Sessions]
      summary: Delete workout session
      operationId: deleteWorkoutSession
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/workout-sessions/{id}/finish:
    post:
      tags: [Workout Sessions]
      summary: Finish (complete) a workout session
      description: |
        Sets `isCompleted = true`. The session upserts by `(firebaseUid, date, name)` —
        finishing an edited session for the same day updates the existing log rather than
        creating a duplicate.
      operationId: finishWorkoutSession
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Completed session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSessionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Plans ────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    FirebaseJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token obtained via client-side Firebase Auth sign-in
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Unauthorized:
      description: Missing or invalid Firebase JWT
    Forbidden:
      description: Authenticated user does not own this resource
    NotFound:
      description: Resource not found
  schemas:
...
    TagEntityType:
      type: string
      enum: [DIET, EXERCISE, MEAL, FOOD]
      description: Which entity type a tag belongs to
...
    TagDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
          nullable: true
          description: Hex or CSS color string (e.g. "#FF5733")
        entityType:
          $ref: '#/components/schemas/TagEntityType'
      required: [id, name, entityType]
...
    FoodDto:
      type: object
      description: |
        A food item. All macros are **per 100 g**. `isSystemFood = true` items are
        shared across all users. `isFavorite` for system foods is stored per-user in
        `food_user_prefs`; for user-owned foods it lives on the food record itself.
      properties:
        id:
          type: integer
          format: int64
          description: Server-assigned DB ID (omit on create)
        serverId:
          type: string
          format: uuid
          nullable: true
          description: Stable client-generated UUID — sync identity
        firebaseUid:
          type: string
          readOnly: true
          description: Owner (set by server from JWT; null for system foods)
          nullable: true
        name:
          type: string
        brand:
          type: string
          nullable: true
        barcode:
          type: string
          nullable: true
        caloriesPer100:
          type: number
          format: double
        proteinPer100:
          type: number
          format: double
        carbsPer100:
          type: number
          format: double
        fatPer100:
          type: number
          format: double
        gramsPerPiece:
          type: number
          format: double
          nullable: true
        gramsPerCup:
          type: number
          format: double
          nullable: true
        gramsPerTbsp:
          type: number
          format: double
          nullable: true
        gramsPerTsp:
          type: number
          format: double
          nullable: true
        glycemicIndex:
          type: integer
          nullable: true
        isSystemFood:
          type: boolean
          default: false
          readOnly: true
        isFavorite:
          type: boolean
          default: false
        verified:
          type: boolean
          default: false
          description: true = from a trusted nutrition database; false = user-created or unverified
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required: [name, caloriesPer100, proteinPer100, carbsPer100, fatPer100]
    FoodPage:
      type: object
      description: Paginated food search result (Spring Page).
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FoodDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
          description: Current page (0-indexed)
        size:
          type: integer
      required: [content, totalElements, totalPages, number, size]

    # ── Meal ─────────────────────────────────────────────────────────────────
...
    ExerciseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        serverId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        isSystem:
          type: boolean
          default: false
          readOnly: true
          description: true = bundled system exercise visible to all users
        tagIds:
          type: array
          items:
            type: integer
            format: int64
          default: []
          description: IDs of EXERCISE-type tags assigned to this exercise
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagDto'
          default: []
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required: [name]

    # ── Workout Templates ─────────────────────────────────────────────────────
    TemplateExerciseDto:
      type: object
      description: An exercise within a workout template with target sets × reps.
      properties:
        id:
          type: integer
          format: int64
        templateId:
          type: integer
          format: int64
        exerciseId:
          type: integer
          format: int64
        orderIndex:
          type: integer
          default: 0
        targetSets:
          type: integer
          default: 3
          minimum: 1
        targetReps:
          type: integer
          nullable: true
          minimum: 1
        targetWeightKg:
          type: number
          format: double
          nullable: true
        notes:
          type: string
          nullable: true
        exerciseName:
          type: string
          readOnly: true
          description: Denormalized exercise name for display convenience
      required: [exerciseId, targetSets]
    WorkoutTemplateDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firebaseUid:
          type: string
          readOnly: true
        name:
          type: string
        notes:
          type: string
          nullable: true
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/TemplateExerciseDto'
          default: []
      required: [name]

    # ── Workout Sessions ──────────────────────────────────────────────────────
    WorkoutSetDto:
      type: object
      description: Actual reps and weight logged for one set of one exercise.
      properties:
        id:
          type: integer
          format: int64
        sessionId:
          type: integer
          format: int64
        exerciseId:
          type: integer
          format: int64
        setNumber:
          type: integer
          description: 0-indexed set order within this exercise in this session
        reps:
          type: integer
          nullable: true
        weightKg:
          type: number
          format: double
          nullable: true
        notes:
          type: string
          nullable: true
      required: [exerciseId, setNumber]
    WorkoutSessionDto:
      type: object
      description: A completed (or in-progress) workout log for one day.
      properties:
        id:
          type: integer
          format: int64
        serverId:
          type: string
          format: uuid
          nullable: true
        firebaseUid:
          type: string
          readOnly: true
        name:
          type: string
          description: Workout / activity name
        date:
          type: string
          format: date
          nullable: true
          description: The day this workout was performed (YYYY-MM-DD)
        durationMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        isCompleted:
          type: boolean
          default: false
        sets:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSetDto'
          default: []
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required: [name]
    LastSetsDto:
      type: object
      description: Sets from the most recent completed session containing the given exercise.
      properties:
        exerciseId:
          type: integer
          format: int64
        sets:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSetDto'
      required: [exerciseId, sets]

    # ── Plans ─────────────────────────────────────────────────────────────────
...